requirements.txt
flask==2.3.2
numpy==1.24.3
tensorflow==2.12.0
pillow==9.5.0
matplotlib==3.7.1
segmentation-models==1.0.1
patchify==0.2.3
scikit-learn==1.2.2
opencv-python==4.7.0.72
pandas==1.3.6
kagglehub==0.1.1
